# openapi/planner-v2.4.yaml
openapi: 3.0.3
info:
  title: APS Planner - Avoidance Decision Model
  description: |
    Microservice wrapper for the APS v2.4 decision model (decision_model.py).

    Position in the xOrbita stack:
    - Upstream: Propagator service supplies relative state and covariance at TCA;
      Ingest supplies optional pre-computed P_c.
    - Downstream: ATLAS presentation layer consumes the machine-readable output;
      Spacecraft GNC receives burn commands.

    Conventions (per Technical Plan):
    - Frame: ECI (TEME internally as needed)
    - Positions: km
    - Velocities: km/s
    - dv magnitudes: m/s
    - Times: UTC, ISO-8601
    - Covariance: km2 (3x3 symmetric, row-major flat array)
  version: 2.4.0
  contact:
    name: xOrbita APS Team

servers:
  - url: http://localhost:8060
    description: Local / Jetson dev
  - url: http://planner:8060
    description: Docker-compose service name

paths:
  /health:
    get:
      operationId: healthCheck
      summary: Liveness probe
      responses:
        "200":
          description: Service is healthy
          content:
            application/json:
              schema:
                type: object
                properties:
                  status:
                    type: string
                    example: ok
                  version:
                    type: string
                    example: "2.4.0"

  /v1/evaluate:
    post:
      operationId: evaluateConjunction
      summary: Evaluate a single conjunction and return the recommended avoidance maneuver
      description: |
        Accepts the relative state, covariance, satellite inertial state,
        and operator policy for one conjunction event. Returns the best
        burn candidate with decision metrics.

        This is the REST equivalent of evaluate_conjunction() in
        avoid/decision_model.py.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/EvaluateRequest"
      responses:
        "200":
          description: Recommended maneuver and decision metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/EvaluateResponse"
        "422":
          description: Validation error (missing/malformed fields)
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

  /v1/evaluate/batch:
    post:
      operationId: evaluateBatch
      summary: Evaluate multiple conjunctions in a single call
      description: |
        Accepts an array of conjunction events. Returns a ranked list of
        recommended maneuvers, one per conjunction, sorted by utility score
        descending.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - conjunctions
              properties:
                conjunctions:
                  type: array
                  items:
                    $ref: "#/components/schemas/EvaluateRequest"
                  minItems: 1
      responses:
        "200":
          description: Array of maneuver recommendations
          content:
            application/json:
              schema:
                type: object
                properties:
                  results:
                    type: array
                    items:
                      $ref: "#/components/schemas/EvaluateResponse"
                  evaluated_at:
                    type: string
                    format: date-time
        "422":
          description: Validation error
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

components:
  schemas:
    EvaluateRequest:
      type: object
      required:
        - conjunction_id
        - satellite
        - conjunction
        - policy
      properties:
        conjunction_id:
          type: string
          description: Unique identifier for this conjunction event
          example: "conj-2026-0302-001"
        satellite:
          $ref: "#/components/schemas/SatelliteState"
        conjunction:
          $ref: "#/components/schemas/ConjunctionState"
        policy:
          $ref: "#/components/schemas/OperatorPolicy"

    SatelliteState:
      type: object
      description: Satellite inertial state at planned burn time
      required:
        - sat_id
        - r_sat_km
        - v_sat_km_s
        - t_burn_utc
        - v_remaining_m_s
      properties:
        sat_id:
          type: string
          description: Satellite identifier
          example: "XORB-001"
        r_sat_km:
          type: array
          description: "Satellite ECI position at burn time [x, y, z] (km)"
          items:
            type: number
          minItems: 3
          maxItems: 3
          example: [6878.0, 0.0, 0.0]
        v_sat_km_s:
          type: array
          description: "Satellite ECI velocity at burn time [vx, vy, vz] (km/s)"
          items:
            type: number
          minItems: 3
          maxItems: 3
          example: [0.0, 7.668, 0.0]
        t_burn_utc:
          type: string
          format: date-time
          description: Planned burn execution time (UTC, ISO-8601)
          example: "2026-03-02T14:00:00Z"
        v_remaining_m_s:
          type: number
          description: Remaining maneuver authority (m/s)
          example: 25.0

    ConjunctionState:
      type: object
      description: Relative geometry at Time of Closest Approach (TCA)
      required:
        - obj_id
        - t_ca_utc
        - r_rel_km
        - p_rel_km2
      properties:
        obj_id:
          type: string
          description: Object / debris identifier
          example: "obj-30337"
        t_ca_utc:
          type: string
          format: date-time
          description: Time of Closest Approach (UTC, ISO-8601)
          example: "2026-03-02T15:30:00Z"
        r_rel_km:
          type: array
          description: "Relative position at TCA [x, y, z] (km). dr = r_obj - r_sat"
          items:
            type: number
          minItems: 3
          maxItems: 3
          example: [0.5, -0.3, 0.1]
        p_rel_km2:
          type: array
          description: |
            Relative covariance at TCA (km2).
            3x3 symmetric matrix in row-major flat form
            [c11, c12, c13, c21, c22, c23, c31, c32, c33].
            P_rel = P_obj + P_sat (assuming independence).
          items:
            type: number
          minItems: 9
          maxItems: 9
          example: [0.01, 0.0, 0.0, 0.0, 0.01, 0.0, 0.0, 0.0, 0.01]
        pc_precomputed:
          type: number
          nullable: true
          description: |
            Optional pre-computed collision probability from ingest.
            If null the planner uses Mahalanobis distance as risk surrogate.
          example: null

    OperatorPolicy:
      type: object
      description: Operator / policy inputs that tune the decision objective
      required:
        - lambda_v
        - lambda_L
        - dv_mag_limit_m_s
      properties:
        lambda_v:
          type: number
          description: Fuel-cost weight in the utility function
          example: 0.01
        lambda_L:
          type: number
          description: Lifetime-penalty weight in the utility function
          example: 0.005
        dv_mag_limit_m_s:
          type: number
          description: Maximum allowable burn magnitude for this event (m/s)
          example: 1.0
        a_ref_km:
          type: number
          description: |
            Reference orbit semi-major axis for Clohessy-Wiltshire STM (km).
            Defaults to 7000.0 km if omitted. Override per-satellite as needed.
          default: 7000.0
          example: 6878.0
        burn_window:
          $ref: "#/components/schemas/BurnWindow"
        hard_constraints:
          $ref: "#/components/schemas/HardConstraints"

    BurnWindow:
      type: object
      description: Temporal constraints on burn execution
      properties:
        earliest_utc:
          type: string
          format: date-time
          description: Earliest permissible burn time (UTC)
        latest_utc:
          type: string
          format: date-time
          description: Latest permissible burn time (UTC)

    HardConstraints:
      type: object
      description: Additional execution constraints
      properties:
        no_burn_windows:
          type: array
          description: Time intervals where burns are forbidden (e.g., eclipse, comms)
          items:
            type: object
            properties:
              start_utc:
                type: string
                format: date-time
              end_utc:
                type: string
                format: date-time
        attitude_restricted:
          type: boolean
          default: false
          description: If true, only along-track burns are permitted
        power_constrained:
          type: boolean
          default: false
          description: If true, apply reduced thrust profile

    EvaluateResponse:
      type: object
      required:
        - conjunction_id
        - recommendation
        - metrics
      properties:
        conjunction_id:
          type: string
          example: "conj-2026-0302-001"
        recommendation:
          $ref: "#/components/schemas/ManeuverRecommendation"
        metrics:
          $ref: "#/components/schemas/DecisionMetrics"
        evaluated_at:
          type: string
          format: date-time
          description: Timestamp when the evaluation was computed

    ManeuverRecommendation:
      type: object
      description: The recommended burn - machine-readable output for GNC / ATLAS
      required:
        - direction
        - dv_eci_km_s
        - dv_magnitude_m_s
        - t_burn_utc
        - utility
      properties:
        direction:
          type: string
          enum: [prograde, radial, cross-track]
          description: Named burn direction (best of the three candidates)
          example: prograde
        dv_eci_km_s:
          type: array
          description: "Recommended dv vector in ECI [dvx, dvy, dvz] (km/s)"
          items:
            type: number
          minItems: 3
          maxItems: 3
          example: [0.0, 0.001, 0.0]
        dv_magnitude_m_s:
          type: number
          description: "Burn magnitude ||dv|| (m/s)"
          example: 1.0
        t_burn_utc:
          type: string
          format: date-time
          description: Scheduled burn time (echoed from request)
        utility:
          type: number
          description: Utility score U = delta_C - lambda_v * ||dv|| - lambda_L * delta_L
          example: 4.23

    DecisionMetrics:
      type: object
      description: Decision-support metrics (machine-readable, consumed by ATLAS for display)
      required:
        - delta_C
        - m2_pre
        - m2_post
        - fuel_cost_m_s
        - lifetime_penalty
      properties:
        delta_C:
          type: number
          description: Confidence gain delta_C = m2_pre - m2_post (Mahalanobis)
          example: 5.12
        m2_pre:
          type: number
          description: Pre-maneuver Mahalanobis squared distance
          example: 8.34
        m2_post:
          type: number
          description: Post-maneuver Mahalanobis squared distance
          example: 3.22
        fuel_cost_m_s:
          type: number
          description: "Fuel expenditure ||dv|| (m/s)"
          example: 1.0
        lifetime_penalty:
          type: number
          description: "Dimensionless lifetime cost = ||dv|| / V_remaining"
          example: 0.04
        risk_surrogate_post:
          type: number
          description: Post-maneuver risk surrogate (1/m2_post or P_c if available)
          example: 0.31
        all_candidates:
          type: array
          description: Full sweep results for all evaluated directions (for audit / ATLAS)
          items:
            type: object
            properties:
              direction:
                type: string
              dv_eci_km_s:
                type: array
                items:
                  type: number
              delta_C:
                type: number
              utility:
                type: number

    ErrorResponse:
      type: object
      properties:
        error:
          type: string
          description: Human-readable error message
          example: "r_rel_km must be a 3-element array"
        detail:
          type: object
          description: Machine-readable validation detail